Crate zoon

Re-exports

pub use dom::document;

pub use dom::history;

pub use dom::load_script;

pub use dom::load_stylesheet;

pub use dom::window;

pub use dominator;

pub use easy_cast;

pub use futures_channel;

pub use futures_signals;

pub use futures_util;

pub use js_sys;

pub use once_cell;

pub use strum;

pub use wasm_bindgen;

pub use wasm_bindgen_futures;

pub use web_sys;

pub use moonlight;

pub use console_error_panic_hook;

pub use web_storage::local_storage;

pub use web_storage::session_storage;

pub use web_storage::LocalStorage;

pub use web_storage::SessionStorage;

pub use web_storage::WebStorage;

pub use serde_json;

pub use serde;

pub use thiserror;

pub use chrono;

Modules

Reflect

ability

button

canvas

checkbox

column

console

dom

ease

el

events

events_extra

future

Asynchronous values.

image

label

link

named_color

Set of colors available for use in Zoon. They can be used in every methods accepting HSLuv colors. The palette is based on https://tailwindcss.com/docs/customizing-colors

oneshot

A channel for sending a single message between asynchronous tasks.

paragraph

raw_el

raw_text

routing

row

signal

spacer

stack

text

text_area

text_input

web_storage

Macros

clone

Macro for cloning values to close. Alternative short record.

element_vec

eprintln

format

Creates a String using interpolation of runtime expressions.

gensym

Generate a unique identifier with a span of Span::call_site and insert it as the first argument to a macro call followed by a comma.

hsluv

make_event

make_flags

map_mut

map_mut is exactly the same as map_ref, except it gives mutable references (map_ref gives immutable references).

map_ref

The map_ref macro can be used to combine multiple Signals together:

paste

println

public_url

run_once

Structs

Align

Styling to align elements inside their container. It is possible to combine different methods to get the styling you need.

AuthToken

Background

Styling to set the background for an element.

Border

Single border definition.

Borders

Styling to set borders for an element.

Broadcaster

Splits an input Signal into multiple output Signals.

Button

Create a customizable Button for your web page. The button is actually built using a div to avoid default behaviors and styling from regular Html buttons. You can create a new button by using its constructor and chain styling.

Canvas

Checkbox

ClassId

Clip

Clip the element by its parent or a mask to show only a specific area.

Closure

A handle to both a closure in Rust as well as JS closure which will invoke the Rust closure.

Column

Connection

CorId

CssPropValue

Cursor

Styling to manage mouse cursor.

Date

ISO 8601 calendar date with time zone.

DateTime

ISO 8601 combined date and time with time zone.

Dom

DomBuilder

Duration

ISO 8601 time duration with nanosecond precision. This also allows for the negative duration; see individual methods for details.

El

EntityId

EventOptions

FixedOffset

The time zone with fixed offset, from UTC-23:59:59 to UTC+23:59:59.

Font

Styling to manage font.

FontLine

Gap

Define gap between elements. More information at https://developer.mozilla.org/en-US/docs/Web/CSS/gap.

GlobalStyles

HSLuv

https://www.hsluv.org/

Height

Styling for height.

Image

IndexGenerator

JsError

Convenience type for use on exported fn() -> Result<T, JsError> functions, where you wish to throw a JavaScript Error object.

JsFuture

A Rust Future backed by a JavaScript Promise.

JsString

JsValue

Representation of an object owned by JS.

KeyboardEvent

Label

LayerIndex

Styling to define the value on the z axis for an element. It does translate to z-index in css.

Link

Local

The local timescale. This is implemented via the standard time crate.

MonotonicIds

MouseEvent

MsgOptions

Mutable

MutableBTreeMap

MutableSignal

MutableSignalMap

MutableSignalVec

MutableVec

NaiveDate

ISO 8601 calendar date without timezone. Allows for every proleptic Gregorian date from Jan 1, 262145 BCE to Dec 31, 262143 CE. Also supports the conversion from ISO 8601 ordinal and week date.

NaiveDateTime

ISO 8601 combined date and time without timezone.

NaiveTime

ISO 8601 time without timezone. Allows for the nanosecond precision and optional leap second representation.

NewTab

Padding

Define padding with pixels for an element.

Paragraph

Placeholder

PointerEvent

PointerHandling

PointerHandlingSvg

RawHtmlEl

RawSvgEl

RawText

ReadOnlyMutable

Receiver

ResizeObserver

RoundedCorners

Define rounded corners. It does translate to css border-radius for web. More information at https://developer.mozilla.org/en-US/docs/Web/CSS/border-radius.

Router

Row

Scene

Scrollbars

Style an element with scrollbars.

SendWrapper

A wrapper which allows you to move around non-Send-types between threads, as long as you access the contained value only from within the original thread and make sure that it is dropped from within the original thread.

Sender

Shadow

Single Shadow effect. More information available at https://developer.mozilla.org/en-US/docs/Web/CSS/box-shadow.

Shadows

Style to add shadows to an element.

SignalStream

SnapAlign

SnapItems

Spacer

Stack

StaticCSSProps

Css properties to be added to the generated html.

StyleAnimation

StyleGroup

Css styles that can be added on a raw html element or globally with a selector.

StyleGroups

StyleRuleHandle

Task

TaskHandle

Text

TextArea

TextContentSelecting

See https://developer.mozilla.org/en-US/docs/Web/CSS/user-select (including Browser compatibility)

TextInput

Timer

TouchHandling

Transform

Define transformation styling to update the shape or position of an element. More information at https://developer.mozilla.org/en-US/docs/Web/CSS/transform.

Transition

Transitions

Define transitions for an element between two states. More information at https://developer.mozilla.org/en-US/docs/Web/CSS/transition

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

Viewport

Visible

Width

Wrapper

Enums

ButtonPressEvent

CursorIcon

Either

ExchangeMsgsError

FontFamily

FontWeight

InputType

Key

Lang

Month

The month of the year.

Radius

Define radius for rounded corners with pixels.

RawElement

RawKeyboardEvent

RawMouseEvent

RawPointerEvent

ReceiveDownMsgError

SecondsFormat

Specific formatting options for seconds. This may be extended in the future, so exhaustive matching in external code is not recommended.

SendUpMsgError

Tag

Weekday

The day of week.

Statics

PUBLIC_URL

VENDOR_PREFIXES

Traits

AddNearbyElement

Also

Represents a type that you can apply arbitrary functions to. Useful for when a method doesn’t return the receiver but you want to apply several of them to the object.

Apply

Represents a type which can have functions applied to it (implemented by default for all types).

ChoosableTag

Datelike

The common set of methods for date component.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

DeserializeOwned

A data structure that can be deserialized without borrowing any data from the deserializer.

DomBuilderExt

Element

FlagNotSet

FlagSet

Focusable

FromRouteSegments

Future

A future represents an asynchronous computation obtained by use of async.

FutureExt

An extension trait for Futures that provides a variety of convenient adapters.

HasIds

HasLang

Hookable

IntoCowStr

IntoDom

IntoEither

IntoElement

IntoOptionCowStr

IntoOptionElement

IntoOptionRadius

JsCast

A trait for checked and unchecked casting between JS types.

KeyboardEventAware

MouseEventAware

MutableViewport

Offset

The offset from the local time to UTC.

PointerEventAware

RawEl

ResizableViewport

RouteSegment

SelectableTextContent

Serialize

A data structure that can be serialized into any data format supported by Serde.

Signal

SignalExt

SignalExtBool

SignalExtExt

SignalExtOption

SignalMap

SignalMapExt

SignalVec

SignalVecExt

StaticEvent

Stream

A stream of values produced asynchronously.

StreamExt

An extension trait for Streams that provides a variety of convenient combinator functions.

Style

Trait to be implemented to enable the use for styling. Every struct such as Align and Background needs to implement this trait so they can be used by Styleable implementations with the s() method within a Zoon element.

Styleable

SubsecRound

Extension trait for subsecond rounding or truncation to a maximum number of digits. Rounding can be used to decrease the error variance when serializing/persisting to lower precision. Truncation is the default behavior in Chrono display formatting. Either can be used to guarantee equality (e.g. for testing) when round-tripping through a lower precision format.

TimeZone

The time zone.

Timelike

The common set of methods for time component.

TouchEventAware

UnwrapThrowExt

An extension trait for Option<T> and Result<T, E> for unwrapping the T value, or throwing a JS error if it is not available.

UpdateRawEl

Functions

always

always_vec

Converts a Vec<A> into a SignalVec<Item = A>

box_css_signal

ch

channel

global_styles

Set styles that are globally used in your application. Very convenient for customizing the design at one single place.

not

pct

public_url

px

start_app

Type Definitions

BoxedCssSignal

DynamicCSSClasses

DynamicCSSProps

StaticCSSClasses

U32Height

U32Width

Attribute Macros

__wasm_bindgen_class_marker

pin_project

An attribute that creates projection types covering all the fields of struct or enum.

route

static_ref

wasm_bindgen

Derive Macros

Deserialize

Serialize