Trait zoon::Style

pub trait Style<'a> where
    Self: Sized, {
    fn move_to_groups(self, groups: &mut StyleGroups<'a>);
}

Trait to be implemented to enable the use for styling. Every struct such as Align and Background needs to implement this trait so they can be used by Styleable implementations with the s() method within a Zoon element.

Required Methods

fn move_to_groups(self, groups: &mut StyleGroups<'a>)