Trait zoon::web_storage::WebStorage

pub trait WebStorage: Sized {
    fn try_new() -> Result<Self>;
    fn storage(&self) -> &Storage;

    fn clear(&self) { ... }
    fn len(&self) -> u32 { ... }
    fn key(&self, index: u32) -> Option<String> { ... }
    fn remove(&self, key: impl AsRef<str>) { ... }
    fn get<T: DeserializeOwned>(&self, key: &str) -> Option<Result<T>> { ... }
    fn insert<T: Serialize + ?Sized>(&self, key: &str, value: &T) -> Result<()> { ... }
}

Web Storage API.

LocalStorage and SessionStorage implement this trait.

MDN reference

Required Methods

fn try_new() -> Result<Self>

Creates a new instance.

Errors

Returns error if we cannot get access to the storage - security errors, browser does not have given storage, user denied access for the current origin, etc.

MDN ref for Local Storage

MDN ref for Session Storage

fn storage(&self) -> &Storage

Get the inner web_sys::Storage instance.

This method is used internally by other methods.

Provided Methods

fn clear(&self)

Clear all data in the storage.

MDN reference

fn len(&self) -> u32

Get the number of stored data items.

MDN reference

fn key(&self, index: u32) -> Option<String>

Returns the key in the given position.

MDN reference

fn remove(&self, key: impl AsRef<str>)

Removes a key.

If there is no item associated with the given key, this method will do nothing.

MDN reference

fn get<T: DeserializeOwned>(&self, key: &str) -> Option<Result<T>>

Returns a deserialized value corresponding to the key.

Errors

Returns error when deserialization fails.

MDN reference

fn insert<T: Serialize + ?Sized>(&self, key: &str, value: &T) -> Result<()>

Insert a key-value pair. The value will be serialized.

If the key already exists, the value will be updated.

Errors

Returns error if we cannot serialize the value or insert/update the pair.

The function web_sys::Storage::set_item is used under the hood. A related warning from MDN docs:

“setItem() may throw an exception if the storage is full. Particularly, in Mobile Safari (since iOS 5) it always throws when the user enters private mode. (Safari sets the quota to 0 bytes in private mode, unlike other browsers, which allow storage in private mode using separate data containers.) Hence developers should make sure to always catch possible exceptions from setItem().”

MDN reference